merged: backported warning fixes from linuxport
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@12435 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ XBMC/xbmc/lib/UnrarXLib/strlist.cpp
@@ -96,13 +96,13 @@ char* StringList::GetString()
 
 bool StringList::GetString(char **Str,wchar **StrW)
 {
-  if (CurPos>=StringData.Size())
+  if ((int)CurPos>=StringData.Size())
   {
     *Str=NULL;
     return(false);
   }
   *Str=&StringData[CurPos];
-  if (PosDataItem<PosDataW.Size() && PosDataW[PosDataItem]==CurPos)
+  if ((int)PosDataItem<PosDataW.Size() && PosDataW[PosDataItem]==(int)CurPos)
   {
     PosDataItem++;
     if (StrW!=NULL)
@@ -119,7 +119,7 @@ bool StringList::GetString(char **Str,wchar **StrW)
 
 char* StringList::GetString(unsigned int StringPos)
 {
-  if (StringPos>=StringData.Size())
+  if ((int)StringPos>=StringData.Size())
     return(NULL);
   return(&StringData[StringPos]);
 }
